refactor types
