+++ src/providers/google/vertex.ts
@@ -18,6 +18,7 @@ import { maybeLoadFromExternalFile } from '../../util';
 import { isValidJson } from '../../util/json';
 import { getNunjucksEngine } from '../../util/templates';
 import { parseChatPrompt, REQUEST_TIMEOUT_MS } from '../shared';
+import type { ClaudeRequest, ClaudeResponse, CompletionOptions, Content } from './types';
 import type { GeminiErrorResponse, GeminiFormat, Palm2ApiResponse } from './util';
 import {
   getGoogleClient,
@@ -26,186 +27,17 @@ import {
   type GeminiResponseData,
 } from './util';
 
-interface Blob {
-  mimeType: string;
-  data: string; // base64-encoded string
-}
-
-interface FunctionCall {
-  name: string;
-  args?: { [key: string]: any };
-}
-
-interface FunctionResponse {
-  name: string;
-  response: { [key: string]: any };
-}
-
-interface FileData {
-  mimeType?: string;
-  fileUri: string;
-}
-
-interface Part {
-  text?: string;
-  inlineData?: Blob;
-  functionCall?: FunctionCall;
-  functionResponse?: FunctionResponse;
-  fileData?: FileData;
-}
-
-interface Content {
-  parts: Part[];
-  role?: string;
-}
-
-interface Schema {
-  type: 'TYPE_UNSPECIFIED' | 'STRING' | 'NUMBER' | 'INTEGER' | 'BOOLEAN' | 'ARRAY' | 'OBJECT';
-  format?: string;
-  description?: string;
-  nullable?: boolean;
-  enum?: string[];
-  maxItems?: string;
-  minItems?: string;
-  properties?: { [key: string]: Schema };
-  required?: string[];
-  propertyOrdering?: string[];
-  items?: Schema;
-}
-
-interface FunctionDeclaration {
-  name: string;
-  description: string;
-  parameters?: Schema;
-  response?: Schema;
-}
-
-export type FunctionParameters = Record<string, unknown>;
-
-interface GoogleSearchRetrieval {
-  dynamicRetrievalConfig: {
-    mode?: 'MODE_UNSPECIFIED' | 'MODE_DYNAMIC';
-    dynamicThreshold?: number;
-  };
-}
-
-interface Tool {
-  functionDeclarations?: FunctionDeclaration[];
-  googleSearchRetrieval?: GoogleSearchRetrieval;
-  codeExecution?: object;
-  googleSearch?: object;
-}
-
-interface VertexCompletionOptions {
-  apiKey?: string;
-  apiHost?: string;
-  projectId?: string;
-  region?: string;
-  publisher?: string;
-  apiVersion?: string;
-  anthropicVersion?: string;
-  anthropic_version?: string; // Alternative format
-
-  // https://ai.google.dev/api/rest/v1beta/models/streamGenerateContent#request-body
-  context?: string;
-  examples?: { input: string; output: string }[];
-  safetySettings?: { category: string; probability: string }[];
-  stopSequence?: string[];
-  temperature?: number;
-  maxOutputTokens?: number;
-  max_tokens?: number; // Alternative format for Claude models
-  topP?: number;
-  top_p?: number; // Alternative format for Claude models
-  topK?: number;
-  top_k?: number; // Alternative format for Claude models
-
-  generationConfig?: {
-    context?: string;
-    examples?: { input: string; output: string }[];
-    stopSequence?: string[];
-    temperature?: number;
-    maxOutputTokens?: number;
-    topP?: number;
-    topK?: number;
-  };
-
-  toolConfig?: {
-    functionCallingConfig?: {
-      mode?: 'MODE_UNSPECIFIED' | 'AUTO' | 'ANY' | 'NONE';
-      allowedFunctionNames?: string[];
-    };
-  };
-
-  tools?: Tool[];
-
-  /**
-   * If set, automatically call these functions when the assistant activates
-   * these function tools.
-   */
-  functionToolCallbacks?: Record<string, (arg: string) => Promise<string>>;
-
-  systemInstruction?: Content;
-
-  /**
-   * Model-specific configuration for Llama models
-   */
-  llamaConfig?: {
-    safetySettings?: {
-      enabled?: boolean;
-      llama_guard_settings?: Record<string, unknown>;
-    };
-  };
-}
-
-// Claude API interfaces
-interface ClaudeMessage {
-  role: string;
-  content: Array<{
-    type: string;
-    text: string;
-  }>;
-}
-
-interface ClaudeRequest {
-  anthropic_version: string;
-  stream: boolean;
-  max_tokens: number;
-  temperature?: number;
-  top_p?: number;
-  top_k?: number;
-  messages: ClaudeMessage[];
-}
-
-interface ClaudeResponse {
-  id: string;
-  type: string;
-  role: string;
-  model: string;
-  content: Array<{
-    type: string;
-    text: string;
-  }>;
-  stop_reason: string;
-  stop_sequence: string | null;
-  usage: {
-    input_tokens: number;
-    cache_creation_input_tokens: number;
-    cache_read_input_tokens: number;
-    output_tokens: number;
-  };
-}
-
 const clone = Clone();
 
 class VertexGenericProvider implements ApiProvider {
   modelName: string;
 
-  config: VertexCompletionOptions;
+  config: CompletionOptions;
   env?: EnvOverrides;
 
   constructor(
     modelName: string,
-    options: { config?: VertexCompletionOptions; id?: string; env?: EnvOverrides } = {},
+    options: { config?: CompletionOptions; id?: string; env?: EnvOverrides } = {},
   ) {
     const { config, id, env } = options;
     this.env = env;
@@ -282,7 +114,7 @@ export class VertexChatProvider extends VertexGenericProvider {
   // https://cloud.google.com/vertex-ai/generative-ai/docs/learn/model-versioning#gemini-model-versions
   constructor(
     modelName: string,
-    options: { config?: VertexCompletionOptions; id?: string; env?: EnvOverrides } = {},
+    options: { config?: CompletionOptions; id?: string; env?: EnvOverrides } = {},
   ) {
     super(modelName, options);
   }
